types: move Result types to requests.ts
